requirements.txt
django-model-utils>=2.0
django-imagekit==4.1.0
Pillow==8.4.0